tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/context.casesensitive.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: context.casesensitive</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>context.casesensitive</h1>
<p class="desc">Context name "2D" is unrecognised; matching is case sensitive</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Context name \"2D\" is unrecognised; matching is case sensitive");
_addTest(function(canvas, ctx) {

_assertSame(canvas.getContext('2D'), null, "canvas.getContext('2D')", "null");


});
</script>
